Phase 79 (HS-79-01): db/activity.py becomes the activity package — done
1,596 lines, one ActivityRepository, eight concerns -> six concern
mixins composed over BaseRepository in a 27-line __init__ (records 348,
settings 215, rules 406, enrichment 248, annotations 138, candidates
303 — all under the Phase-63 module budget with headroom). The public
surface is unchanged; db.activity call sites untouched.

Executed by line-range slicing, verbatim: a programmatic check against
git show HEAD: found zero method-body lines differing; the nine
non-verbatim lines are import plumbing only (distributed model imports,
the three lazy activity_mapping imports one package level deeper,
datetime into records' header). Tests unmodified — zero patch-target
edits: unit 2407 passed, integration 685 passed.
