+++ pm/roadmap/holdspeak/phase-79-backend-decomposition-ii/current-phase-status.md
@@ -19,7 +19,7 @@ guard-locked so nothing regrows.
 
 | ID | Title | Status |
 |----|-------|--------|
-| HS-79-01 | `db/activity.py` → the activity package (concern mixins) | todo — leads |
+| HS-79-01 | `db/activity.py` → the activity package (concern mixins) | **done** ([evidence](./evidence-story-01.md): 1,596 → six mixins, largest 406; zero body drift; tests unmodified, 2407 + 685 green) |
 | HS-79-02 | `routes/system.py` → the system package (five routers) | todo |
 | HS-79-03 | `routes/primitives.py` → the primitives package (per family + shared run helpers) | todo |
 | HS-79-04 | The density guard locks the new shapes | todo |
@@ -28,7 +28,7 @@ guard-locked so nothing regrows.
 
 ## Where we are
 
-Scaffolded. The carve order is 01 → 02 → 03 (largest first; all three independent), then
+**01 done** (the activity package). The carve order continues 02 → 03 (largest first; all three independent), then
 the guard, docs, closeout. One PR on `phase-79-backend-decomposition-ii`, merged on
 conclusion-checked green.
